Clustering/requirements.txt
faerun==0.4.7
matplotlib==3.9.4
numpy==2.3.3
pandas==2.3.3
Pillow==11.3.0
scikit_learn==1.6.1
tmap==1.0.6
